Improve Caddy support, add git config
This makes it easy to automagically add support for a domain to Caddy,
as well as remove/replace it if need be. It's not perfect (you still
have to remember what port your service is going to run on), but it will
do for now.

Additionally, add a bunch of git config stuff. This moves over all of my
old git config, as well as adding some new things.